backend/backend_eng/requirements.txt
# Backend Engineering Dependencies
# Owner: Backend Engineering Team

# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.0.0

# Configuration
python-dotenv>=1.0.0